refactor(S-3.01): extract handle_status to auth/status.rs (BC-7.4.014)
Move peek_oauth_app_source, peek_oauth_app_source_for_test, and status()
out of mod.rs. peek_oauth_app_source_for_test re-exported under #[cfg(test)]
for BC-1.4.030 holdout tests. BC-7.4.014 auth state inspection preserved.
612 unit tests pass.
